scrapers/requirements.txt
celery[redis]==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
fake-useragent==1.4.0
python-dateutil==2.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-dotenv==1.0.0